+++ src/app/locales/en/components/-storage/ceph.js
@@ -5,6 +5,4 @@ export default {
   clusterName: { name: 'Cluster name' },
   poolName: { name: 'Pool name' },
   timeout: { name: 'Timeout [ms]' },
-  insecure: { name: 'Insecure' },
-  readonly: { name: 'Read only' },
 };