+++ src/app/locales/en/components/-storage/glusterfs.js
@@ -6,6 +6,4 @@ export default {
   mountPoint: { name: 'Relative mountpoint in volume' },
   xlatorOptions: { name: 'Custom client translator options' },
   timeout: { name: 'Timeout [ms]' },
-  insecure: { name: 'Insecure' },
-  readonly: { name: 'Read only' },
 };