+++ src/app/locales/en/components/-storage/nulldevice.js
@@ -39,6 +39,4 @@ export default {
       'while 100.0 will increase it by 100 files per second.',
   },
   timeout: { name: 'Timeout [ms]' },
-  insecure: { name: 'Insecure' },
-  readonly: { name: 'Read only' },
 };