+++ src/app/locales/en/components/-storage/posix.js
@@ -1,6 +1,4 @@
 export default {
   mountPoint: { name: 'Mount point' },
   timeout: { name: 'Timeout [ms]' },
-  readonly: { name: 'Read only' },
-  insecure: { name: 'Insecure' },
 };