+++ src/app/locales/en/components/-storage/swift.js
@@ -6,6 +6,4 @@ export default {
   tenantName: { name: 'Tenant name' },
   blockSize: { name: 'Block size' },
   timeout: { name: 'Timeout [ms]' },
-  insecure: { name: 'Insecure' },
-  readonly: { name: 'Read only' },
 };