+++ src/app/locales/en/components/-storage/webdav.js
@@ -21,7 +21,7 @@ export default {
   },
   onedataAccessToken: {
     name: 'Onedata access token',
-    tip: 'When registering storage in "insecure" mode with "OAuth2" external IdP, this field must contain a valid Onedata access token of the user on whose behalf the WebDAV storage will be accessed by all users with access to any space supported by this storage.',
+    tip: 'When registering storage with feed of LUMA DB set to "auto" and with "OAuth2" external IdP, this field must contain a valid Onedata access token of the user on whose behalf the WebDAV storage will be accessed by all users with access to any space supported by this storage.',
   },
   authorizationHeader: {
     name: 'Authorization header',
@@ -41,16 +41,11 @@ export default {
   },
   fileMode: {
     name: 'File mode',
-    tip: 'Defines the file permissions, which files imported from WebDAV storage will have in Onedata. Values should be provided in octal format e.g. "0644". Default: 0644.',
+    tip: 'Defines the file permissions, which files imported from WebDAV storage will have in Onedata. Values should be provided in octal format e.g. "0664". Default: 0664.',
   },
   dirMode: {
     name: 'Directory mode',
     tip: 'Defines the directory mode which directories imported from WebDAV storage will have in Onedata. Values should be provided in octal format e.g. "0775".  Default: 0775.',
   },
   timeout: { name: 'Timeout [ms]' },
-  insecure: {
-    name: 'Insecure',
-    tip: 'Defines whether storage administrator credentials (username and key) may be used by users without storage accounts to access storage in direct IO mode.',
-  },
-  readonly: { name: 'Read only' },
 };