+++ src/app/utils/cluster-storage/ceph-fields.js
@@ -5,6 +5,4 @@ export default [
   { name: 'clusterName', type: 'text' },
   { name: 'poolName', type: 'text' },
   { name: 'timeout', type: 'number', optional: true },
-  { name: 'insecure', type: 'checkbox', optional: true },
-  { name: 'readonly', type: 'checkbox', optional: true },
 ];