+++ src/app/utils/cluster-storage/posix-fields.js
@@ -1,5 +1,4 @@
 export default [
   { name: 'mountPoint', type: 'text' },
   { name: 'timeout', type: 'number', optional: true },
-  { name: 'readonly', type: 'checkbox', optional: true },
 ];