+++ src/lib/ember-onedata-onepanel-server/addon/utils/cluster-storage-class.js
@@ -27,6 +27,10 @@ const {
   GlusterfsModify,
   Webdav,
   WebdavModify,
+  HTTP,
+  HTTPModify,
+  XRootD,
+  XRootDModify,
   Nulldevice,
   NulldeviceModify,
 } = Onepanel;
@@ -53,8 +57,12 @@ function clusterStorageClass(storageType, modify = false) {
       return modify ? SwiftModify : Swift;
     case 'glusterfs':
       return modify ? GlusterfsModify : Glusterfs;
+    case 'xrootd':
+      return modify ? XRootDModify : XRootD;
     case 'webdav':
       return modify ? WebdavModify : Webdav;
+    case 'http':
+      return modify ? HTTPModify : HTTP;
     case 'nulldevice':
       return modify ? NulldeviceModify : Nulldevice;
     default: